Three changes the LASTRADA DBTable→Lightweight migration needs from the
query and variant layers. Bundled so the migration takes one
overlay-port bump, not three.

| Commit | What | Why the migration needs it |
|---|---|---|
| SELECT/DELETE bound inputs | `Select`/`Delete` take a bindings vector
like `Insert`/`Update` already do | A criteria list assembled at run
time (length and type mix known only then) can bind by marker instead of
keeping markers and values in step by hand. Ergonomics, not a new
capability — `SqlWildcard` + a prepared statement already bound a
SELECT's WHERE. |
| JOIN ON clause: values, null, groups | `OnValue`/`OrOnValue`,
`OnNull`/`OnNotNull`/`OrOn…`, `OnGroup`/`OrOnGroup` beside `On`/`OrOn` |
The BO selectors put constants and null tests inside outer-join ON
clauses (a tenant filter, an archive flag). Moving such a term to WHERE
turns a LEFT OUTER JOIN into an inner one. |
| Unsigned TINYINT read faithfully | `SQL_TINYINT` reads into `short`,
not signed `int8_t`, in both the per-row and prefetch paths | SQL Server
TINYINT is unsigned 0..255. A value >127 made the per-row path fail the
row with 22003 — and, inside the `noexcept` cursor, **crash** — while
prefetch silently wrapped 255→−1. Any table with a LOB column (disables
prefetch) plus a TINYINT>127 crashed on read. |

**Verification:** full suite green on SQLite (1599/1) and MS SQL Server
2022 (1597/0); the crash reproduced against a real production row before
the fix and passes after.

Bound-ON values insert ahead of WHERE values (JOIN precedes WHERE in the
statement); `SqlSearchCondition::joinBindingCount` keeps the vector in
statement order regardless of call interleaving.

---
**Known limitation (out of scope, unexercised):**
`SqlUpdateQueryBuilder::ToSql()` does not emit `tableJoins`, so an
`UPDATE().Join(...).OnValue(...)` would bind a value with no matching
marker. UPDATE+JOIN was already a silent no-op before this PR; no caller
does it. A guard belongs with a future proper UPDATE-with-JOIN
implementation (dialect-specific syntax).

+++ src/Lightweight/SqlStatement.cpp
@@ -1002,7 +1002,9 @@ SqlVariant SqlStatement::MakePrefetchVariantCell(RowArrayCursor const& cursor, s
         case SQL_BIT:
             return SqlVariant { SqlVariant::InnerType { static_cast<bool>(cursor.GetI64(row, column).value_or(0)) } };
         case SQL_TINYINT:
-            return SqlVariant { SqlVariant::InnerType { static_cast<int8_t>(cursor.GetI64(row, column).value_or(0)) } };
+            // Unsigned 0..255 on SQL Server; int8_t would wrap 128..255 to negatives. Widen to short,
+            // matching the per-row SqlVariant path so a prefetched cell compares equal to a per-row one.
+            return SqlVariant { SqlVariant::InnerType { static_cast<short>(cursor.GetI64(row, column).value_or(0)) } };
         case SQL_SMALLINT:
             return SqlVariant { SqlVariant::InnerType { static_cast<short>(cursor.GetI64(row, column).value_or(0)) } };
         case SQL_INTEGER:
